--- src/components/footer/Siwtch.scss
@@ -1,68 +1,0 @@
-$break-small: 500px;
-$break-large: 1024px;
-
-@mixin respond-to($media) {
-  @if $media == handhelds {
-    @media only screen and (max-width: $break-small) {
-      @content;
-    }
-  } @else if $media == medium-screens {
-    @media only screen and (min-width: $break-small + 1) and (max-width: $break-large - 1) {
-      @content;
-    }
-  } @else if $media == wide-screens {
-    @media only screen and (min-width: $break-large) {
-      @content;
-    }
-  }
-}
-
-.react-switch-checkbox {
-  height: 0;
-  width: 0;
-  visibility: hidden;
-}
-
-.react-switch-label {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-  width: 100px;
-  height: 50px;
-  background: #777fa3;
-  border-radius: 100px;
-  position: relative;
-  transition: background-color 0.2s;
-  @include respond-to(medium-screens) {
-    transform: scale(0.8);
-  }
-  @include respond-to(handhelds) {
-    transform: scale(0.5);
-  }
-}
-
-.react-switch-label .react-switch-button {
-  content: "";
-  position: absolute;
-  top: 2px;
-  left: 2px;
-  width: 45px;
-  height: 45px;
-  border-radius: 45px;
-  transition: 0.2s;
-  background: #fff;
-  box-shadow: 0 0 2px 0 rgba(10, 10, 10, 0.29);
-  @include respond-to(medium-screens) {
-    transform: scale(0.8);
-  }
-}
-
-.react-switch-checkbox:checked + .react-switch-label .react-switch-button {
-  left: calc(100% - 2px);
-  transform: translateX(-100%);
-}
-
-.react-switch-label:active .react-switch-button {
-  width: 60px;
-}
